🔨 Relocate insertNewUser to userQueries
