+++ server/userQueries.ts
@@ -10,3 +10,18 @@ export const findUserByEmail = async (email: string) => {
     console.log('findUserByEmail error', error)
   }
 }
+
+export const insertNewUser = async (email: string, hashedPassword: string) => {
+  const newUser = {
+    email: email,
+    password: hashedPassword,
+  }
+  try {
+    const { usersCollection, close } = await dbConnect()
+    const dbRes = await usersCollection.insertOne(newUser)
+    close()
+    return dbRes
+  } catch (error) {
+    console.log('insertNewUser error', error)
+  }
+}
